requirements.txt
# Core dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# Encryption
cryptography>=41.0.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# Quantization
bitsandbytes>=0.41.0
peft>=0.6.0
accelerate>=0.24.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0

# Evaluation
scikit-learn>=1.3.0

# Logging
colorlog>=6.7.0
